### 2021-04-09 #3163 @whedon
[PRE REVIEW]: FIRESONG: A python package to simulate populations of extragalactic neutrino sources
**Submitting author:** @ChrisCFTung (<a href="http://orcid.org/0000-0001-6920-7841">Chun Fai Tung</a>)
**Repository:** <a href="https://github.com/icecube/FIRESONG" target ="_blank">https://github.com/icecube/FIRESONG</a>
**Version:** v1.6
