### 2021-04-14 c4785f4
be consistent with diffuse flux assumed
