+++ firesong/FluxPDF.py
@@ -24,8 +24,8 @@ def flux_pdf(outputdir,
              zmin=0.0005,
              zmax=10.,
              bins=10000,
-             fluxnorm=0.9e-8,
-             index=2.13,
+             fluxnorm=1.44e-8,
+             index=2.28,
              LF="SC",
              sigma=1.0,
              luminosity=0.0,
