### 2021-04-20 #3194 @JostMigenda
Hi @ChrisCFTung, thanks for the submission! I’ll install & test the software in a couple of days; for now, just a few initial comments and questions, after looking at the repo and the paper:

* Comparing the author list with Whedon’s software report, I’ve noticed that Theo Glauch doesn’t appear to have contributed any code. Is that an inaccuracy in the git history? Or can you confirm that one of JOSS’ other [authorship criteria](https://joss.readthedocs.io/en/latest/submitting.html#authorship) is fulfilled instead?
* In the paper, the Summary and Background sections are good. However, the following sections are currently not as clear as they could be:
  * In the Statement of Need, the first sentence doesn’t actually say what FIRESONG does. Can you add 1–2 sentences that explicitly state this? (For example, this could be something like: "FIRESONG can be used to generate source distributions based on a theoretical model selected by the user and calculate the neutrino flux arriving on Earth, taking into account theoretical predictions for initial flux, propagation effects like redshift, etc.”)
  * Also in the Statement of Need, could you add references for [Numpy](https://numpy.org/citing-numpy/) and [SciPy](https://www.scipy.org/citing.html)?
  * For a reader who is not yet familiar with FIRESONG, I think the first paragraph of the Usage section is a bit overwhelming. In part, I think this is related to my previous bullet point—if I don’t know what exactly FIRESONG does, it’s hard to put all the options you list here into context.
  * For the second paragraph on the Usage section, can you add an explanation of what Legend does and in what situations you would use it instead of Firesong? (You did this quite well in the third paragraph, for FluxPDF.)
* The version strings in your [__init__.py](https://github.com/icecube/FIRESONG/blob/base/firesong/__init__.py#L5) and [setup.py](https://github.com/icecube/FIRESONG/blob/base/setup.py#L4) don’t match the version tags in the repo, which may be confusing.
* In your README file, can you state what Python version is required?
