### 2021-04-22 c6ca5d4
consistent version numbers in setup, include python version in setup
