### 2021-04-23 #3194 @ChrisCFTung
Hi @JostMigenda,

Thank you for your comments and questions. Firstly regarding Theo Glauch's contribution, the git history was lost during a migration from Georgia Institute of Technology's git repository to github. In particular, `FluxPDF.py` was contributed by Theo.

We have also addressed the other issues you mentioned.

Regarding the comments on the paper:
1. Description of FIRESONG's usage has been added in the [Statement of Need](https://github.com/icecube/FIRESONG/blob/base/paper/paper.md#statement-of-need) section.
2. Added refereces for Numpy and Scipy
3. Added description for LEGEND in [USAGE](https://github.com/icecube/FIRESONG/blob/base/paper/paper.md#usage), to specify that LEGEND should be used when users want to simulate sources with a distribution that depends on luminosity.

Regarding [__init__.py](https://github.com/icecube/FIRESONG/blob/base/firesong/__init__.py#L5) and [setup.py](https://github.com/icecube/FIRESONG/blob/base/setup.py#L4), version string has been updated.

Regarding the README file, Python3 is now specified in [Set-Up](https://github.com/icecube/FIRESONG/blob/base/README.md#set-up).
