### 2021-04-26 icecube/FIRESONG#55 @JostMigenda
Installation instructions
_[As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/3194).]_

I ran into several issues while trying to install FIRESONG.
- [x] The installation instructions in the README file are currently incomplete: If you recommend installing FIRESONG as an editable install from a local folder (`pip install --editable FIRESONG`), you need to state explicitly that the user should download the repo first.
However, note that edible installs are mainly for developers of a package. For users, publishing FIRESONG to PyPI and installing via `pip install FIRESONG` (i.e. without downloading the repo) may be preferable; both for simplicity and to ensure they install a tagged release (as opposed to the most recent git commit on the base branch, which is more prone to cause confusion).
- [x] Right now, installing FIRESONG does not install any dependencies; so after installing it into an empty conda environment, I get an error message ("ModuleNotFoundError: No module named 'numpy'") when trying to run e.g. `from firesong.Firesong import firesong_simulation`. You can use setup.py’s [`install_requires` keyword](https://packaging.python.org/guides/distributing-packages-using-setuptools/#install-requires) to let pip install these dependencies automatically.
- [x] Since FIRESONG documentation is mainly in the form of Jupyter notebooks, it would be useful to add jupyter and matplotlib as dependencies.
