### 2021-04-26 icecube/FIRESONG#55 @JostMigenda
Additionally, one of the dependencies (CosmoloPy) apparently fails to install under Python 3.8 or newer. I’ve [opened an issue](https://github.com/roban/CosmoloPy/issues/9) on the CosmoloPy repo.
Please let me know whether there’s a workaround or whether I should continue testing with Python 3.7 for now.
