### 2021-04-28 f331935
1. added pypi installation to readme 2. added install_requirement to setup.py
