### 2021-04-29 7676e6f
This update removed the dependecy on cosmolopy 1. all cosmology-related calculations are handled by firesong/distance.py 2. ```SourcePopulation.cosmology``` can now provide distance calculation 3. slight change to test_evolution.py, observed differences:    standard candle flux --> 9.244753729666578e-11 vs 9.244753729666597e-11    Nsources  --> 3745.1796382799967  vs  3745.179638278482    transient redshift distribution --> difference at 4 decimal place
