+++ firesong/distance.py
@@ -1,0 +1,183 @@
+#!/usr/bin/python
+
+'''
+Distance calculation substitute for cosmolopy
+All formula used are from 
+https://arxiv.org/pdf/astro-ph/9905116.pdf
+
+Default in all functions are using Planck15
+'''
+import numpy as np
+from scipy.integrate import quad
+
+class cosmo_distance(object):
+    def __init__(self, **cosmology):
+        self.c = 299792458 # speed of light in ms^-1
+        # initialize class
+        self.load_param(**cosmology)
+
+    def load_param(self, **cosmology):
+        '''
+        set up the cosmological parameters
+        density of curvature is defined as 1-om0-ode
+        unless otherwise specified.
+        '''
+        self.om0 = cosmology.get('omega_M_0', 0.315)
+        self.ode = cosmology.get('omega_lambda_0', 0.685)
+        self.ok0 = cosmology.get('omega_k_0', 1-self.om0-self.ode)
+        self.h = cosmology.get('h', 0.674)
+
+        # Hubble distance D_h (unit=Mpc)
+        self.D_h = self.c/1000./100./self.h
+
+    def E(self, z):
+        '''
+        time derivative of the logarithm of the scale factor
+        E(z) = \frac{\dot{a}(t)}{a(t)}
+
+        Input:
+        z     : float or array of float, redshift
+        kwargs:  'omega_M_0'     
+                 'omega_lambda_0'
+                 'omega_k_0'
+                 density parameters
+        output:
+        E      : float or array of float
+        '''
+        EE = np.sqrt(self.om0*(1.+z)**3.+self.ok0*(1.+z)**2.+self.ode)
+        return EE
+
+    def comoving_distance(self, z, z_0=0):
+        '''
+        line-of-sight comoving distance
+        D_c = D_h \int^z_0 \frac{dz'}{E(z')}
+        unit = Mpc
+        Input:
+        z     : float or array of float, redshift
+        z0    : float, default is 0, where integration begins
+        kwargs:  'omega_M_0'     
+                 'omega_lambda_0'
+                 'omega_k_0'
+                 density parameters
+        output:
+        D_c      : float or array of float
+
+        '''
+        z = np.atleast_1d(z)
+        D_c = np.array([self.D_h*quad(lambda x:1./self.E(x), 
+                                 z_0, lim, epsabs=1.e-5)[0] for lim in z])
+        # return a float if only one redshift
+        if np.size(D_c) > 1:
+            return D_c
+        else:
+            return D_c[0]
+
+    def transverse_comoving_distance(self, z):
+        '''
+        The comoving distance between two events at the same redshift or distance 
+        but separated on the sky by some angle is D_m * d\theta
+        D_m is the transverse comoving distance
+        if omega_R > 0:
+            D_m = D_h /\sqrt(Omega_R) \sinh(\sqrt(Omega_R)*D_c/D_h)
+        if omega_R = 0:
+            D_m = D_c
+        if omega_R < 0:
+            D_m = D_h /\sqrt(\abs(Omega_R)) \sin(\sqrt(\abs(Omega_R))*D_c/D_h)
+
+        unit = Mpc
+
+        Input:
+        z     : float or array of float, redshift
+        kwargs:  'omega_M_0'     
+                 'omega_lambda_0'
+                 'omega_k_0'
+                 density parameters
+        output:
+        D_m      : float or array of float
+        '''
+        if self.ok0==0:
+            return self.comoving_distance(z)
+        elif self.ok0>0:
+            return self.D_h/np.sqrt(self.ok0)*np.sinh(np.sqrt(self.ok0)\
[…]
