+++ firesong/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = 'C.F. Tung, T. Glauch, M. Larson, A. Pizzuto, R. Reimann, I. Taboada'
 __email__ = ''
-__version__ = '1.6.0'
+__version__ = '1.6.1'
