+++ requirements.txt
@@ -1,4 +1,3 @@
-CosmoloPy==0.4
 coverage==5.4
 numpy==1.16.6
 scipy==1.2.3
