### 2021-04-30 icecube/FIRESONG#55 @JostMigenda
[This commit](https://github.com/icecube/FIRESONG/commit/f331935d34dee2ad423b9d95050d07079bf867a2) solves the issues listed above, thank you! And in the meantime, the CosmoloPy issue was also fixed and it can now be pip-installed under Python 3.8 as well, so I will continue to use that for testing.

3.9 isn’t supported by CosmoloPy yet, though; it would probably be good to mention that in the README file to save users some trouble trying to debug it.
Also, the version you uploaded to PyPI is still missing the jupyter and matplotlib requirements, unfortunately; it looks like you uploaded an incomplete version of the code there?

PS: If you weren’t already aware of it, check out [this GitHub Action](https://help.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions#publishing-to-package-registries) to automate publishing to PyPI whenever you create a new release; maybe that’s helpful to you?
