### 2021-04-30 e210949
1. cosmo_distance class now check the input 2. change jupyter requirement in setup.py
