+++ setup.py
@@ -24,5 +24,5 @@ setuptools.setup(
     'numpy>=1.16.6',
     'scipy>=1.2.3',
     'matplotlib',
-    'jupyter']
+    'jupyter>=3.0']
 )
