### 2021-04-30 e44f4f6
changed jupyter and matplotlib dependencies
