+++ requirements.txt
@@ -1,5 +1,3 @@
 coverage==5.4
 numpy==1.16.6
 scipy==1.2.3
-matplotlib==3.2.2
-jupyter==4.6.3
