### 2021-04-30 icecube/FIRESONG#59 @JostMigenda
Interaction between fluxnorm and zmin
_[As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/3194).]_

Here’s a usage example:
```
from firesong.Firesong import firesong_simulation

sim = firesong_simulation(None, filename=None,
                                density=1e-10,
                                Evolution='MD2014SFR',
                                zmin=0.0005,  # <-- Default value
                                verbose=False)
print("=== zmin = 0.0005 ===")
print(f"Expected flux: {sim['header']['fluxnorm']}")
print(f"Actual flux of {len(sim['sources']['flux'])} sources: {sim['total_flux']}")

sim = firesong_simulation(None, filename=None,
                                density=1e-10,
                                Evolution='MD2014SFR',
                                zmin=4,  # <-- Note this non-default value!
                                verbose=False)
print("=== zmin = 4 ===")
print(f"Expected flux: {sim['header']['fluxnorm']}")
print(f"Actual flux of {len(sim['sources']['flux'])} sources: {sim['total_flux']}")
```
And this is the output I got:
> === zmin = 0.0005 ===
> Expected flux: 1.44e-08
> Actual flux of 1329 sources: 1.528777753275981e-08
> === zmin = 4 ===
> Expected flux: 1.44e-08
> Actual flux of 1329 sources: 5.704090191660857e-10

The actual flux will vary randomly but is usually between 1e-8 and 2e-8 for zmin=-0.0005 (comparable to the expected flux normalization) while it is between 5.5e-10 and 6e-10 for zmin=4 (i.e. more than 20 times smaller than the expected flux normalization). The number of sources is constant across runs.

I would have expected one of two things:
1. The flux normalization describes the flux from sources in the selected redshift range, between zmin and zmax. (This would imply that the actual flux should always be close to the flux normalization.)
2. The flux normalization describes the flux integrated over all redshifts (starting at 0.0), but the code outputs only the fraction of sources that are between zmin and zmax. (This would imply that selecting a smaller z range should result in fewer sources being generated.)

Neither option appears to be the case. Am I misunderstanding something or is this a bug?
