### 2021-04-30 eddc074
Expanded example in README
