### 2021-04-30 icecube/FIRESONG#55 @ChrisCFTung
Hi @JostMigenda,

Sorry about the confusion, that previous version I uploaded to pypi was not ready. In order to address the issues you mentioned, I have released a new version v.1.6.1 (also updated on PyPI), which no longer require cosmolopy. I believe this can avoid any future python version issue. The switch has effectively no impact on the result.

regarding the requirement of `jupyter` and `matplotlib`, I think they are not really involved in the core functions of FIRESONG, so I don't want to force users to install these packages. However, I have stated explicitly in the README that access to the examples requires jupyter notebook and matplotlib. Please let me know if that make sense to you.

And thank you for the suggestion on publishing automatically to PyPI. I will definitely implement that.
