### 2021-04-30 #3194 @rafaelab
Hi @ChrisCFTung,

I have read the paper and I am now done testing the software.
It looks fine overall, easy to install and easy to use. The examples are very helpful.
I will divide my comments in two parts: first about the paper, second about the code.

--> PAPER

Overall I find the paper well-written and I have only minor comments.

- FIRESONG was motivated by IceCube, but it clearly has many applications other than neutrino astronomy. For instance, I had to write down a similar code for ultra-high-energy cosmic rays (UHECRs), which turns out to be much slower (~30 times slower) than FIRESONG. It is as useful for neutrinos as it is for UHECRs (and gamma rays). Perhaps these other applications could be briefly mentioned in the paper to reach a broader audience.

- The paper refers to cosmolopy, which has now been removed as a dependency, so it should be updated accordingly.

--> CODE

I followed the tutorials and looked into the code to try it out. I had no problems installing nor running it.
My comments are more about things that could be useful and some minor suggestions.

- Maybe this is me being picky, but I found it annoying that the keyword "Evolution" in ```firesong_simulation()``` is capitalised when all the other keywords are not. I got this wrong twice when trying to run a simple test without copy-pasting. The same goes for "Transient".

- When running ```firesong_simulation()```, I get a dictionary that contains 'sources'. This is also a dictionary containing 'flux', 'z', 'dec'. I tried to write down a function that takes the dictionary and plots the skymap automatically, then I realised 'ra' was missing. I do understand the reason for not including it in the code at first. But I think it would be a useful addition to have this stored in the dictionary for plotting skymaps without having to write down more lines of code. This is very useful for anisotropy analyses.

- I am also missing an example with Legend in the notebooks folder.
