### 2021-05-03 icecube/FIRESONG#55 @JostMigenda
Nice to see that removing the dependency on cosmolopy was possible without too much trouble; this should indeed prevent most issues with future Python versions. Regarding `jupyter` and `matplotlib`, that seems like a good compromise to me, yes. Thanks!
