### 2021-05-03 icecube/FIRESONG#59 @JostMigenda
To add to this: When I vary `zmax` instead of `zmin`, the number of sources does change as I would expect in the second case. So it looks like number 2 is the correct option and there is simply a bug related to `zmin`?
