### 2021-05-03 #3194 @JostMigenda
I have now finished testing the software and have opened issues on the FIRESONG repository where appropriate. These are mostly minor points that shouldn’t take long to resolve, so I hope that we will reach acceptance of this paper before too long.

Apart from the noted issues, the code works well and seems both versatile and useful. The notebooks give good examples for how to use the code and plot the results; they also nicely illustrate how Firesong and FluxPDF relate to each other. (Like Rafael, I would welcome it if you could add some examples for how to use Legend, too.)

One final point that’s missing right now—as far as I can see— is this one:
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Just referring people to PRs/Issues on GitHub is perfectly fine, but this should be stated explicitly somewhere; probably in the README file.

Related to community guidelines: I see that the IceCube collaboration has a [code of conduct](https://res.cloudinary.com/icecube/images/v1600803899/icc-governance-v8.8/icc-governance-v8.8.pdf) (see Appendix F). Since the FIRESONG repository is under the IceCube organisation on GitHub, I’d assume that it applies to all interactions in this repo as well? If so, that would be good to mention as well; right next to the contribution guidelines.
