### 2021-05-03 icecube/FIRESONG#59 @ChrisCFTung
Hi @JostMigenda,

Sorry for the late reply. Actually zmin was not supposed to be an option for the user. The reason is that the neutrino flux generated depends on the local density of the source, so the minimum distance to the source can not be at zmin.  It is therefore a bug.

I am working on an update that will remove this option.
