### 2021-05-03 ca5d83f
Updates to address listed issues 1. removed `zmin` from Firesong, FluxPDF, and Legend 2. adjusted README for clarity 3. added ra to output for Firesong and Legend 4. added notebook for legend
