### 2021-05-03 icecube/FIRESONG#64 @ChrisCFTung
1. Doc string of `flux_pdf` has been addressed in #65
2. Indeed the inconsistency of the options is an issue for user experience, but considering the impact that can be brought by changing those, I think it's a better choice to keep it the same
