### 2021-05-03 icecube/FIRESONG#63 @ChrisCFTung
The requirement for the folder `Results` has been removed in #65 .
Both example scripts should work now.
In README, I also specify the location the output file, so users should be able to find it easily
