### 2021-05-04 icecube/FIRESONG#64 @JostMigenda
That's fair. Of course, if there's ever a FIRESONG v2.0 which breaks backwards compatibility anyway, feel free to revisit this …
