### 2021-05-04 #3194 @itaboada
Interesting! The governance document (including the code of conduct (appendix F) are visible publicly! It's not like that we (IceCube) were hiding the document ...
I will need to consult with others if the code of conduct, which was developed for IceCube internal affairs, can or should be applied to non-IceCube people.
