### 2021-05-04 a68f032
fixed typo in input_output also switched to `os.path.joint` to handle file name
