### 2021-05-04 icecube/FIRESONG#63 @ChrisCFTung
`Results` folder is no longer need in the output directory, so the example can be executed even if `FIRESONG` is not set.
I have also added in README the location of the output file, which should help user to locate the output.
