### 2021-05-04 icecube/FIRESONG#62 @ChrisCFTung
Changed the example in README, and the way to handle filename in the newest commit
