### 2021-05-04 icecube/FIRESONG#59 @ChrisCFTung
zmin has been removed in #65
