### 2021-05-04 #3194 @JostMigenda
Yes—it's linked from IceCube's public web page at https://icecube.wisc.edu/collaboration/institutions/

I noticed that the "Scope" section of the CoC already mentioned the case of "an event participant that is not an IceCube collaborator"; that's why it seemed plausible to me that it could apply here.
I understand that this is not what the CoC was originally designed for; on the other hand, having a separate CoC and separate reporting structures may introduce unnecessary confusion, especially since any parties to potential conflicts may likely be IceCube members?

Either way—thanks @itaboada for starting that discussion internally! As far as I understand (@eloisabentivegna, correct me if I'm wrong!), JOSS doesn't require a CoC so while it would be good to clarify this, it shouldn't stop us from proceeding with the review.
