### 2021-05-04 #3194 @ChrisCFTung
Thank you for the comment @rafaelab .

Regarding your comment on the paper, in the paper, we have mentioned publications which applied Firesong on studies of gamma rays in the statement of need section of the paper. Indeed, providing a function to convert Firesong's output to other messengers is one of the open issue for Firesong (#45). However, before that service is mature, we don't want to propose a "standard" way of conversion to the users in this paper, because such conversion is heavily model dependent. We are leaving this up to the users' own capabilities for now.

Regarding the comments for the code
1. the options will not be changed for the moment, due to possible breakage of existing scripts. However, we will look for a safe way to introduce more coherent option in the future.

2. Added RA to the output, which is sampled from a uniform distribution. Since we expect the extragalactic sources to be isotropic

3. Added a notebook as example for Legend.
