### 2021-05-04 52aa541
Corrected the name for model HA2012BL
