### 2021-05-08 #3194 @rafaelab
Hi @ChrisCFTung,

> 1. the options will not be changed for the moment, due to possible breakage of existing scripts. However, we will look for a safe way to introduce more coherent option in the future.

I completely understand that this breaks previous scripts. I saw @JostMigenda's issue and he recommended what I was going to suggest: that these conventions are kept in mind for future releases, if there are any.

> 2. Added RA to the output, which is sampled from a uniform distribution. Since we expect the extragalactic sources to be isotropic

I am aware this is fairly simple, but I think it makes the code more complete and enables easy manipulations of the results as a dictionary.

> 3. Added a notebook as example for Legend.

Having successfully tested this, I have no further comments and I am happy to recommend the paper for publication.
