### 2021-05-09 #3194 @eloisabentivegna
> (@eloisabentivegna, correct me if I'm wrong!), JOSS doesn't require a CoC so while it would be good to clarify this, it shouldn't stop us from proceeding with the review.

You are correct, @JostMigenda.
