### 2021-05-09 #3194 @eloisabentivegna
Thanks @JostMigenda and @rafaelab for your prompt and thorough reviews. Loads of good points raised and improves. Thanks also to @ChrisCFTung and @itaboada for being so responsive.

I am happy to proceed with prepublication. @ChrisCFTung, could you generate a tagged version and submission archive, ad per https://joss.readthedocs.io/en/latest/submitting.html#the-review-process? Please post the details here once you are done.
